% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse}
\alias{collapse}
\title{Collapse a \code{visreg_list} into a single \code{visreg} object}
\usage{
collapse(x, labels)
}
\arguments{
\item{x}{A \code{visreg_list} object.}

\item{labels}{A character vector with length corresponding to the number of \code{visreg} objects in
\code{x} that provides labels for the different objects in subsequent plots. If missing, the
response variable names are used (or \code{Y1}, \code{Y2}, ... if these are not unique).}
}
\value{
A \code{visreg} object.
}
\description{
Takes a \code{visreg_list}, as produced by \code{\link[=visreg_list]{visreg_list()}} or \code{\link[=visreg]{visreg()}}, and collapses it down into
a single \code{visreg} object. This is equivalent to calling \code{\link[=visreg_list]{visreg_list()}} with \code{collapse = TRUE}
in the first place; it is provided separately so that a \code{visreg_list} can be collapsed after
the fact, e.g. \code{visreg_list(v1, v2) |> collapse()}.
}
\examples{
fit <- lm(Ozone ~ Solar.R + Wind + Temp, data = airquality)
v1 <- visreg(fit, "Wind", plot = FALSE, alpha = 0.2)
v2 <- visreg(fit, "Wind", plot = FALSE, alpha = 0.01)
labs <- c("Confidence: 0.80", "Confidence: 0.99")
vv1 <- visreg_list(v1, v2, collapse = FALSE) |> collapse(labels = labs)
vv2 <- visreg_list(v1, v2, collapse = TRUE, labels = labs)
identical(vv1, vv2)

}
\seealso{
\code{\link[=visreg_list]{visreg_list()}}, \code{\link[=visreg]{visreg()}}
}
