% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persp-visreg2d.R, R/plot-visreg_list.R
\name{persp.visreg2d}
\alias{persp.visreg2d}
\alias{persp.visreg_list}
\title{Static 3D perspective plot for a visreg2d object}
\usage{
\method{persp}{visreg2d}(
  x,
  xlab = NULL,
  ylab = NULL,
  zlab = NULL,
  color = "#2fa4e7",
  whitespace = 0.2,
  ...
)

\method{persp}{visreg_list}(x, ...)
}
\arguments{
\item{x}{A \code{\link[=visreg2d]{visreg2d()}} object.}

\item{xlab}{Axis label for x variable}

\item{ylab}{Axis label for y variable}

\item{zlab}{Axis label for outcome}

\item{color}{The color of the surface.}

\item{whitespace}{When \code{xvar} or \code{yvar} is a factor, \code{whitespace} determines the amount of space
in between the factors. Default is 0.2, meaning that 20 percent of the axis is whitespace.}

\item{...}{Additional graphical parameters, passed to \code{\link[graphics:persp]{graphics::persp()}}.}
}
\value{
The viewing transformation matrix, invisibly; see \code{\link[graphics:persp]{graphics::persp()}}.
}
\description{
\code{persp()} method for \code{\link[=visreg2d]{visreg2d()}} objects: draws a static 3-dimensional perspective plot of
the fitted surface, using base R's \code{\link[graphics:persp]{graphics::persp()}}.
}
\details{
For an interactive version that can be rotated with the mouse, see \code{rgl::persp3d()}
(requires the \code{rgl} package). For a 2-dimensional raster/contour plot, see \code{\link[=plot.visreg2d]{plot.visreg2d()}}.
}
\examples{
fit <- lm(
  Ozone ~ Solar.R + Wind + Temp + I(Wind^2) + I(Temp^2) +
    I(Wind * Temp) + I(Wind * Temp^2) + I(Temp * Wind^2) + I(Temp^2 * Wind^2),
  data = airquality
)

v <- visreg2d(fit, x = "Wind", y = "Temp", plot = FALSE)
persp(v)

}
\references{
Breheny P and Burchett W. (2017) Visualization of regression models using
visreg. \emph{R Journal}, \strong{9}: 56-71.
\doi{10.32614/RJ-2017-046}
}
\seealso{
\code{\link[=visreg2d]{visreg2d()}} for creating two-dimensional \code{visreg} objects, \code{\link[=plot.visreg2d]{plot.visreg2d()}} for the
2-dimensional raster/contour plot, and the
\href{https://pbreheny.github.io/visreg/articles/surface.html}{surface plots vignette} for examples
and details.
}
