% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-visreg2d.R
\name{plot.visreg2d}
\alias{plot.visreg2d}
\title{Visualization of regression functions for two variables}
\usage{
\method{plot}{visreg2d}(
  x,
  xlab = NULL,
  ylab = NULL,
  zlab = NULL,
  color,
  print_cond = FALSE,
  whitespace = 0.2,
  plot.type = NULL,
  ...
)
}
\arguments{
\item{x}{A \code{\link[=visreg2d]{visreg2d()}} object.}

\item{xlab}{Axis label for x variable}

\item{ylab}{Axis label for y variable}

\item{zlab}{Label for the color legend}

\item{color}{A vector of colors used to establish the color palette for the fill/legend.}

\item{print_cond}{If \code{print_cond==TRUE}, the explanatory variable values conditioned on in a
conditional plot are printed to the console (default: \code{FALSE}). If \code{print_cond==TRUE} and
\code{type=="contrast"}, the conditions will still be printed, but they have no bearing on the plot
unless interactions are present.}

\item{whitespace}{When \code{xvar} or \code{yvar} is a factor, \code{whitespace} determines the amount of space
in between the factors. Default is 0.2, meaning that 20 percent of the axis is whitespace.}

\item{plot.type}{Deprecated, no longer has any effect. See
\href{https://pbreheny.github.io/visreg/articles/surface.html}{vignette}.}

\item{...}{Not used.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot method for visualizing how two variables interact to affect the response in regression
models, as a \code{ggplot2} raster/contour plot.
}
\details{
For 3-dimensional surface plots, see \code{\link[=persp.visreg2d]{persp.visreg2d()}} (static) or \code{rgl::persp3d()}
(interactive, requires the \code{rgl} package).
}
\examples{
fit <- lm(
  Ozone ~ Solar.R + Wind + Temp + I(Wind^2) + I(Temp^2) +
    I(Wind * Temp) + I(Wind * Temp^2) + I(Temp * Wind^2) + I(Temp^2 * Wind^2),
  data = airquality
)

visreg2d(fit, x = "Wind", y = "Temp")
visreg2d(fit, x = "Wind", y = "Temp", color = c("purple", "green", "red"))

}
\references{
Breheny P and Burchett W. (2017) Visualization of regression models using
visreg. \emph{R Journal}, \strong{9}: 56-71.
\doi{10.32614/RJ-2017-046}
}
\seealso{
\code{\link[=visreg2d]{visreg2d()}} for creating two-dimensional \code{visreg} objects, \code{\link[=persp.visreg2d]{persp.visreg2d()}} for static 3D
surface plots, and the
\href{https://pbreheny.github.io/visreg/articles/surface.html}{surface plots vignette} for examples
and details.
}
