% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-visreg.R, R/plot-visreg_list.R
\name{plot.visreg}
\alias{plot.visreg}
\alias{plot.visreg_list}
\title{Visualization of regression functions}
\usage{
\method{plot}{visreg}(
  x,
  overlay = FALSE,
  print_cond = FALSE,
  partial = identical(x$meta$trans, I),
  band = TRUE,
  rug = ifelse(partial, 0, 2),
  strip_names = is.numeric(x$fit[, x$meta$by]),
  top = c("line", "points"),
  line = NULL,
  fill = NULL,
  points = NULL,
  gg = TRUE,
  ...
)

\method{plot}{visreg_list}(x, ...)
}
\arguments{
\item{x}{A \code{visreg} or \code{visreg_list} object; see \code{\link[=visreg]{visreg()}}.}

\item{overlay}{By default, when \code{by} is specified, separate panels are used to display each
cross-section. If \code{overlay=TRUE}, these cross-sections are overlaid on top of each other in a
single plot.}

\item{print_cond}{If \code{print_cond=TRUE}, the explanatory variable values conditioned on in a
conditional plot are printed to the console (default: \code{FALSE}). If \code{print_cond=TRUE} and
\code{type="contrast"}, the conditions will still be printed, but they have no bearing on the plot
unless interactions are present.}

\item{partial}{If \code{partial=TRUE} (the default), partial residuals are shown on the plot.}

\item{band}{If \code{band=TRUE} (the default), confidence bands are shown on the plot.}

\item{rug}{By default, partial residuals are plotted. Alternatively, a rug may be plotted along
the horizontal axis instead. Setting \code{rug=TRUE} turns off partial residuals by default; if one
wants both to be plotted, both \code{rug=TRUE} and \code{partial=TRUE} need to be specified. Two types of
rug plots are available. If \code{rug=1} or \code{rug=TRUE}, then a basic rug is drawn on the bottom. If
\code{rug=2}, then separate rugs are drawn on the top for observations with positive residuals and
on the bottom for observations with negative residuals. Such plots are particularly useful in
logistic regression (see examples).}

\item{strip_names}{When \code{by=TRUE}, \code{strip_names=TRUE} adds the name of the \code{by} variable to the
strip at the top of each panel. Default is \code{FALSE} for factors and \code{TRUE} for numeric \code{by}
variables. \code{strip_names} can also be a character vector, in which case it replaces the strip
names altogether with values chosen by the user.}

\item{top}{By default, the fitted line is plotted on top of the partial residuals; usually this
is preferable, but it does run the risk of obscuring certain residuals. To change this behavior
and plot the partial residuals on top, specify \code{top='points'}.}

\item{line}{List of parameters to pass to \code{\link[ggplot2:geom_line]{ggplot2::geom_line()}} (continuous \code{xvar}) or
\code{\link[ggplot2:geom_crossbar]{ggplot2::geom_crossbar()}} (factor \code{xvar}) when the fitted line is plotted.}

\item{fill}{List of parameters to pass to \code{\link[ggplot2:geom_ribbon]{ggplot2::geom_ribbon()}} (continuous \code{xvar}) or
\code{\link[ggplot2:geom_crossbar]{ggplot2::geom_crossbar()}} (factor \code{xvar}) when shaded confidence regions are plotted.}

\item{points}{List of parameters to pass to \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}} (continuous \code{xvar}) or
\code{\link[ggplot2:geom_jitter]{ggplot2::geom_jitter()}} (factor \code{xvar}) when partial residuals are plotted.}

\item{gg}{Deprecated, no longer has any effect.}

\item{...}{Not used; present only because \code{plot()} is a generic function. An error is raised if
any arguments are passed here.}
}
\value{
A ggplot object, or if \code{x} is a \code{\link[=visreg_list]{visreg_list()}}, a list of ggplot objects.
}
\description{
Default plots contain a confidence band, prediction line, and partial residuals. Factors,
transformations, conditioning, interactions, and a variety of other options are supported. The
\code{plot.visreg()} function accepts a \code{visreg} object as calculated by \code{\link[=visreg]{visreg()}} and creates the
plot.
}
\details{
If \code{x} is a \code{\link[=visreg_list]{visreg_list()}}, \code{plot.visreg_list()} is dispatched instead; it simply calls
\code{plot.visreg()} on each element in turn, passing along the same arguments to each element.
}
\examples{
fit <- lm(Ozone ~ Solar.R + Wind + Temp, data = airquality)
visreg(fit, "Wind", line = list(color = "red"), points = list(size = 1, shape = 1))

# Changing appearance
visreg(fit, "Wind", line = list(color = "red"), points = list(size = 1, shape = 1))

}
\references{
Breheny P and Burchett W. (2017) Visualization of regression models using
visreg. \emph{R Journal}, \strong{9}: 56-71.
\doi{10.32614/RJ-2017-046}
}
\seealso{
\link{visreg2d} for creating \code{visreg} objects, and the
\href{https://pbreheny.github.io/visreg/surface.html}{options vignette} for examples.
}
