% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visreg-package.R
\docType{package}
\name{visreg-package}
\alias{visreg-package}
\title{visreg: Visualization of Regression Models}
\description{
Provides a convenient interface for constructing plots to visualize the fit of regression models arising from a wide variety of models in R ('lm', 'glm', 'coxph', 'rlm', 'gam', 'locfit', 'lmer', 'randomForest', etc.)
}
\examples{
\donttest{vignette("quick-start", package="visreg")}
}
\references{
Breheny P and Burchett W. (2017) Visualization of regression models using
visreg. \emph{R Journal}, \strong{9}: 56-71.
\doi{10.32614/RJ-2017-046}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pbreheny.github.io/visreg}
  \item \url{https://github.com/pbreheny/visreg}
  \item Report bugs at \url{https://github.com/pbreheny/visreg/issues}
}

}
\author{
\strong{Maintainer}: Patrick Breheny \email{patrick-breheny@uiowa.edu} (\href{https://orcid.org/0000-0002-0650-1119}{ORCID})

Other contributors:
\itemize{
  \item Woodrow Burchett [contributor]
}

}
\keyword{internal}
