\name{visreg-package}
\alias{visreg-package}
\docType{package}
\title{Visualization of regression functions}
\description{\code{visreg} provides a number of plotting functions for
  visualizing fitted regression models: regression functions, confidence
  bands, partial residuals, interactions, and more.}
\details{
\tabular{ll}{
Package: \tab visreg\cr
Type: \tab Package\cr
Version: \tab 1.0-0\cr
Date: \tab 2011-08-09\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{Patrick Breheny and Woodrow Burchett}
Maintainer: Patrick Breheny <patrick.breheny@uky.edu>}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
